+++ coordinator/store/harness_test.go
@@ -60,6 +60,7 @@ func testPostgresStore(t testing.TB) *PostgresStore {
 		"provider_floor_draws",
 		"code_attestations",
 		"code_attest_push_budgets",
+		"app_attest_build_qualifications",
 		"app_attest_shadow_keys",
 		"app_attest_enrollments",
 		"app_attest_receipts",